requirements.txt
scipion-pyworkflow
scipion-em
scipion-chem

setuptools~=58.0.4